requirements.txt
numpy
streamlit
python-dotenv
google-genrativeai
